Moved neoletterInstance.js into utils folder
